+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/item/domain/dessert/DessertItemCategory.java
@@ -1,6 +1,7 @@
-package git_kkalnane.backend.starbucks.item.domain;
+package git_kkalnane.backend.starbucks.item.domain.dessert;
 
 import git_kkalnane.backend.starbucks._global.entity.BaseTimeEntity;
+import git_kkalnane.backend.starbucks.item.domain.ItemCategory;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
